@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from .cache_manager import CacheManager
-from .repository import Repository
-
+from .cache_manager import CacheManager  # noqa: F401
+from .repository import Repository  # noqa: F401
 
 __version__ = "0.3.0"
