@@ -35,4 +35,4 @@ class MsgPackSerializer(Serializer):
 
         :rtype: str
         """
-        return msgpack.unpackb(data, encoding='utf-8')
+        return msgpack.unpackb(data, encoding="utf-8")
