@@ -2,17 +2,17 @@
 
 from functools import partial
 
+from .serializer import Serializer
+
 try:
-    import cPickle as pickle
-except ImportError:  # noqa
+    import cPickle as pickle  # noqa: N813
+except ImportError:
     import pickle
 
 # Serialize pickle dumps using the highest pickle protocol (binary, default
 # uses ascii)
 dumps = partial(pickle.dumps, protocol=pickle.HIGHEST_PROTOCOL)
 loads = pickle.loads
-
-from .serializer import Serializer
 
 
 class PickleSerializer(Serializer):
