@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
-from .dict_store import DictStore
-from .file_store import FileStore
-from .memcached_store import MemcachedStore
-from .redis_store import RedisStore
-from .null_store import NullStore
+from .dict_store import DictStore  # noqa: F401
+from .file_store import FileStore  # noqa: F401
+from .memcached_store import MemcachedStore  # noqa: F401
+from .null_store import NullStore  # noqa: F401
+from .redis_store import RedisStore  # noqa: F401
